added common function for display name
